temp save: Refractor command definition out from index.js
