--- commands.js
@@ -1,11 +1,0 @@
-export default const commands = [
-  // define slash commands
-  {
-    name: 'ping',
-    description: 'Replies with Pong!',
-  },
-  {
-    name: 'temp',
-    description: 'temp command',
-  },
-];
