+++ commands/ping.js
@@ -1,0 +1,9 @@
+export default {
+  data: {
+    name: 'ping',
+    description: 'Replies with pong!',
+  },
+  async execute(interaction) {
+    await interaction.reply('Pong from ping.js!');
+  },
+};
